Use dropbox/rules_node to build JS packages
 - Load jquery, backbone, and underscore from cdnjs instead of hosting
   locally (and removes them from the local repo)
 - Use webpack (via bazel) to build a JS bundle for each page

This removes all vendored third-party js libraries (except for html.js;
I couldn't find that in NPM, and anyway it's not that big).

JS source is organized under web/src/, and the road is paved to
refactor it into files/modules.

The major drawback is that now `livegrep` must be run with -docroot set
to a bazel-built location (I've been passing `-docroot
bazel-bin/cmd/livegrep/livegrep.runfiles/com_github_livegrep_livegrep/web/`),
and JS changes require a rebuild (`bazel build //web:webpack_build`
is usually good enough).
