+++ web/3d/BUILD
@@ -1,0 +1,9 @@
+package(default_visibility = ['//visibility:public'])
+
+load('@org_dropbox_rules_node//node:defs.bzl', 'node_internal_module')
+
+node_internal_module(
+    name = 'html',
+    srcs = ['html.js'],
+    main = 'html.js',
+)
