--- web/htdocs/assets/js/3d/jquery.js
@@ -1,1 +1,0 @@
-jquery-1.7.1.min.js
