--- web/htdocs/assets/js/3d/underscore.js
@@ -1,1 +1,0 @@
-underscore-min.js
