+++ web/templates/fileview.html
@@ -62,11 +62,3 @@
     </div>
   </section>
 </section>
-
-<script language="javascript" type="text/javascript">
-  // Distasteful. Someday pass this to the fileview code more cleanly.
-  window.repoInfo = {{.Repo}};
-  window.commit = {{.Commit}};
-</script>
-<script language="javascript" type="text/javascript" src="/assets/js/3d/highlight.pack.js"></script>
-<script language="javascript" type="text/javascript" src="/assets/js/fileview.js"></script>
