+++ web/templates/index.html
@@ -1,10 +1,3 @@
-<script type="text/javascript">
-  $(function(){
-    CodesearchUI.repo_urls = {{.RepoUrls}};
-    CodesearchUI.intervalViewRepos = {{.InternalViewRepos}};
-    CodesearchUI.onload();
-  });
-</script>
 <div id='searcharea'>
   <div class="search-inputs">
     <div class="prefixed-input filter-code">
