+++ web/templates/layout.html
@@ -3,15 +3,22 @@
 <html xmlns="http://www.w3.org/1999/xhtml">
   <head>
     <title>{{.Title}}</title>
-    <link rel="stylesheet" href='/assets/css/codesearch.css' />
-    <link rel="stylesheet" href="/assets/css/3d/hljs/tomorrow.css" />
     <link rel="shortcut icon" href="/assets/img/favicon.ico" />
-    <script language="javascript" type="text/javascript" src="/assets/js/3d/jquery.js"></script>
-    <script language="javascript" type="text/javascript" src="/assets/js/3d/underscore.js"></script>
-    <script language="javascript" type="text/javascript" src="/assets/js/3d/backbone.js"></script>
-    <script language="javascript" type="text/javascript" src="/assets/js/3d/html.js"></script>
-    <script language="javascript" type="text/javascript" src="/assets/js/codesearch.js"></script>
-    <script language="javascript" type="text/javascript" src="/assets/js/codesearch_ui.js"></script>
+    <link rel="stylesheet" href='/assets/css/codesearch.css' />
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/1.7.1/jquery.min.js" integrity="sha256-iBcUE/x23aI6syuqF7EeT/+JFBxjPs5zeFJEXxumwb0=" crossorigin="anonymous"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/underscore.js/1.3.1/underscore-min.js" integrity="sha256-Qtj60TvCj8cmd1GW7Jq5U/6/m94XXFhFEoNhyVP6F/Q=" crossorigin="anonymous"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/backbone.js/0.9.2/backbone-min.js" integrity="sha256-tQjdUhE0MTzHcOzRUuotgnMrURWIamfdqwv1QWB57uk=" crossorigin="anonymous"></script>
+    {{if .ScriptData}}
+    <script>
+      window.scriptData = {{.ScriptData}};
+    </script>
+    {{end}}
+    {{if .ScriptName}}
+    <script>
+      window.page = {{.ScriptName}};
+    </script>
+    {{end}}
+    <script language="javascript" type="text/javascript" src="/assets/js/bundle.js"></script>
     {{if .Config.Sentry.URI}}
     <script src="//cdn.ravenjs.com/1.1.15/jquery,native/raven.min.js"></script>
     <script>
